show a QR code
